move InviteMember to a separate file
